+++ qmi/instruments/wl_photonics/__init__.py
@@ -1,0 +1,6 @@
+"""WL Photonics, wavelength-tunable filter.
+
+The qmi.instruments.wl_photonics package provides support for:
+  - WLTF-N: Narrowband USB wavelength-tunable filters.
+"""
+from qmi.instruments.wl_photonics.wltf_n import WlPhotonics_WltfN